// src/enums/IssueStatus.ts
export enum IssueStatus {
  INDEFINIDA = "-",
  A_FAZER = "A FAZER",
  EM_REFINAMENTO = "EM REFINAMENTO",
  EM_EXECUCAO = "EM EXECUÇÃO",
  EM_EXPLORACAO = "EM EXPLORAÇÃO",
  EM_VALIDACAO_DE_HIPOTESES = "EM VALIDAÇÃO DE HIPÓTESES",
  EM_TESTES = "EM TESTES",
  EM_VALIDACAO = "EM VALIDAÇÃO",
  VALIDADO = "VALIDADO",
  EM_IMPLANTACAO = "EM IMPLANTAÇÃO",
  IMPLANTADO = "IMPLANTADO",
  ENCERRADO = "ENCERRADO",
  ATIVADO = "ATIVADO",
  ENCERRADO_ATIVADO = "ENCERRADO/ATIVADO",
  CANCELADO = "CANCELADO",
}